c backend update

+++ tests/c/apl.h
@@ -9,8 +9,8 @@
 
 #define max(x,y) (((x) > (y)) ? (x) : (y))
 #define min(x,y) (((x) < (y)) ? (x) : (y))
-#define i2d(x) (x)
-#define d2i(x) (x)
+#define i2d(x) ((double)x)
+#define d2i(x) ((int)x)
 #define b2i(x) ((x)?1:0)
 #define true 1
 #define false 0
